omit error prefix, print executed command because of levensthein, rename commands
