+++ generate.js
@@ -5,19 +5,15 @@ const {
   writeFile,
   existsSync
 } = require('fs')
-const chalk = require('chalk')
 const path = require('path')
+const chalk = require('chalk')
 const generify = require('generify')
 const argv = require('yargs-parser')
 const cliPkg = require('./package')
 const { execSync } = require('child_process')
+const log = require('./log')
 
-function generate (dir, log, cb) {
-  if (!cb) {
-    cb = log
-    log = () => {}
-  }
-
+function generate (dir, cb) {
   generify(path.join(__dirname, 'templates', 'app'), dir, {}, function (file) {
     log('debug', `generated ${file}`)
   }, function (err) {
@@ -78,20 +74,10 @@ function generate (dir, log, cb) {
   })
 }
 
-const levels = {
-  'debug': 0,
-  'info': 1,
-  'error': 2
-}
-
-const colors = [
-  (l) => l,
-  chalk.green,
-  chalk.red
-]
-
 function cli (args) {
   const opts = argv(args)
+
+  log('info', 'Execute command: gen-app')
 
   if (opts._[0] && existsSync(opts._[0])) {
     log('error', 'directory ' + opts._[0] + ' already exists')
@@ -105,20 +91,12 @@ function cli (args) {
     process.exit(1)
   }
 
-  generate(dir, log, function (err) {
+  generate(dir, function (err) {
     if (err) {
       log('error', err.message)
       process.exit(1)
     }
   })
-
-  function log (severity, line) {
-    const level = levels[severity] || 0
-    if (level === 1) {
-      line = '--> ' + line
-    }
-    console.log(colors[level](line))
-  }
 }
 
 module.exports = {
