+++ help/gen-app.txt
