+++ help/gen-readme.txt
@@ -1,0 +1,5 @@
+Usage: fastify gen-readme <path-to-plugin-file>
+
+Sets up project with `npm init -y` and
+generates a sample README file
+from the <path-to-plugin-file>.
