--- help/generate-plugin.txt
@@ -1,1 +1,0 @@
-Usage: fastify generate-plugin <file>
