+++ help/help.txt
@@ -1,9 +1,9 @@
 Fastify command line interface available commands are:
 
-  * start             start a server
-  * generate          generate a new project
-  * generate-plugin   generate the plugin README.md
-  * version           the current fastify-cli version
-  * help              help about commands
+  * start           start a server
+  * gen-app         generate a new project
+  * gen-readme      generate a README.md for the plugin
+  * version         the current fastify-cli version
+  * help            help about commands
 
 Launch 'fastify help [command]' to learn more about each command.
